# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelOwnReservationDto": {
        "properties": {
          "reason": {
            "description": "Optional. Free text, for the venue.",
            "maxLength": 280,
            "type": "string"
          }
        },
        "type": "object"
      },
      "CancelReservationDto": {
        "properties": {
          "reason": {
            "description": "REQUIRED. Shown to the diner verbatim.",
            "example": "Burst pipe in the kitchen — we are so sorry.",
            "maxLength": 300,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
      "CancelledReservationResponse": {
        "properties": {
          "cancel_reason": {
            "description": "Shown to the diner verbatim.",
            "type": "string"
          },
          "cancelled_at": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "party_size": {
            "type": "integer"
          },
          "starts_at": {
            "type": "string"
          },
          "status": {
            "description": "Always `cancelled_by_restaurant`.",
            "type": "string"
          },
          "table_released": {
            "description": "True when the table this booking held is now available again. Released by trg_resv_propagate, which flips reservation_tables.active off for any status outside held|pending|confirmed|seated.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "cancelled_at",
          "cancel_reason",
          "starts_at",
          "party_size",
          "table_released"
        ],
        "type": "object"
      },
...
      "ImageResponse": {
        "properties": {
          "height": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "is_cover": {
            "description": "The venue hero. Exactly one per owner.",
            "type": "boolean"
          },
          "position": {
            "type": "integer"
          },
          "urls": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Width in px → public URL. Pick the smallest that fits.",
            "example": {
              "160": "https://…/160.webp",
              "400": "…",
              "1200": "…"
            },
            "type": "object"
          },
          "width": {
            "description": "The ORIGINAL's width, for the aspect box.",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "urls",
          "width",
          "height",
          "position",
          "is_cover"
        ],
        "type": "object"
      },
      "JoinWaitlistDto": {
        "properties": {
          "desiredDate": {
            "description": "The VENUE'S wall-clock day.",
            "example": "2026-08-12",
            "type": "string"
          },
          "partySize": {
            "maximum": 50,
            "minimum": 1,
            "type": "integer"
          },
          "restaurantId": {
            "format": "uuid",
            "type": "string"
          },
          "windowEnd": {
            "example": "2026-08-12T19:00:00.000Z",
            "type": "string"
          },
          "windowStart": {
            "description": "Absolute instant.",
            "example": "2026-08-12T16:00:00.000Z",
            "type": "string"
          }
        },
        "required": [
          "restaurantId",
          "desiredDate",
          "windowStart",
          "windowEnd",
          "partySize"
        ],
        "type": "object"
      },
...
      "ModifyReservationDto": {
        "properties": {
          "partySize": {
            "example": 4,
            "maximum": 50,
            "minimum": 1,
            "type": "integer"
          },
          "startsAt": {
            "description": "The new start, ISO-8601 UTC. Must be in the future.",
            "example": "2026-08-09T17:00:00.000Z",
            "type": "string"
          }
        },
        "type": "object"
      },
      "MyReservationResponse": {
        "properties": {
          "can_review": {
            "description": "Whether POST /reviews would accept this visit right now — seated or completed, the table time is over, and not already reviewed. DERIVED ON THE SERVER from the same function the endpoint enforces, so the control the client draws cannot disagree with the answer it gets.",
            "type": "boolean"
          },
          "cancel_reason": {
            "nullable": true,
            "type": "string"
          },
          "cancelled_at": {
            "nullable": true,
            "type": "string"
          },
          "cancelled_by": {
            "description": "'user' | 'restaurant' | null. Derived, so no client parses a status string.",
            "nullable": true,
            "type": "string"
          },
          "code": {
            "description": "Human-readable, quoted at the door. e.g. SAH-7K2M",
            "type": "string"
          },
          "date": {
            "description": "YYYY-MM-DD on the RESTAURANT'S wall clock.",
            "type": "string"
          },
          "ends_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "needs_acknowledgement": {
            "description": "The RESTAURANT cancelled and the diner has not seen it yet. THE CLIENT MUST SURFACE THIS. It is the only signal that a booking they believe they hold is gone, and a reservation carrying it stays in `upcoming` regardless of date until POST /reservations/{id}/acknowledge-cancellation.",
            "type": "boolean"
          },
          "occasion": {
            "nullable": true,
            "type": "string"
          },
          "party_size": {
            "type": "integer"
          },
          "restaurant": {
            "$ref": "#/components/schemas/ReservationVenueResponse"
          },
          "review_id": {
            "description": "The review this visit already has. Null is the common case.",
            "nullable": true,
            "type": "string"
          },
          "source": {
            "description": "app | walk_in | phone — which door it came through.",
            "type": "string"
          },
          "special_requests": {
            "nullable": true,
            "type": "string"
          },
          "starts_at": {
            "description": "Absolute instant, ISO-8601 UTC.",
            "type": "string"
          },
          "status": {
            "description": "pending | confirmed | seated | completed | no_show | cancelled_*",
            "type": "string"
          },
          "time": {
            "description": "HH:MM on the RESTAURANT'S wall clock.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "source",
          "starts_at",
          "ends_at",
          "date",
          "time",
          "party_size",
          "needs_acknowledgement",
          "can_review",
          "restaurant"
        ],
        "type": "object"
      },
...
      "ReservationVenueResponse": {
        "properties": {
          "city": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "timezone": {
            "description": "IANA zone `date` and `time` are expressed in.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name_en",
          "name_ar",
          "city",
          "timezone"
        ],
        "type": "object"
      },
...
      "SaveVenueDto": {
        "properties": {
          "restaurantId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "restaurantId"
        ],
        "type": "object"
      },
      "SavedVenueResponse": {
        "properties": {
          "city": {
            "type": "string"
          },
          "cover": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageResponse"
              }
            ],
            "nullable": true
          },
          "cuisines": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "price_band": {
            "nullable": true,
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "rating_count": {
            "type": "integer"
          },
          "saved_at": {
            "description": "When it was saved. Drives the newest-first order.",
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name_en",
          "name_ar",
          "cuisines",
          "city",
          "rating",
          "rating_count",
          "saved_at"
        ],
        "type": "object"
      },
...
      "WaitlistEntryResponse": {
        "properties": {
          "desired_date": {
            "description": "The VENUE'S wall-clock day, YYYY-MM-DD.",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "offer_expires_at": {
            "description": "Set only while `offered`. C-3.6 gives a 10-minute claim window, and a CHECK constraint ties the two together in both directions.",
            "nullable": true,
            "type": "string"
          },
          "party_size": {
            "type": "integer"
          },
          "restaurant": {
            "$ref": "#/components/schemas/WaitlistVenueResponse"
          },
          "status": {
            "description": "waiting | offered | converted | expired | cancelled",
            "type": "string"
          },
          "window_end": {
            "description": "And the latest. An offer outside this range is worse than none.",
            "type": "string"
          },
          "window_start": {
            "description": "Absolute instant. The earliest the diner will accept.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "desired_date",
          "window_start",
          "window_end",
          "party_size",
          "restaurant"
        ],
        "type": "object"
      },
      "WaitlistVenueResponse": {
        "properties": {
          "city": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name_en",
          "name_ar",
          "city"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Restaurant reservation platform — Egypt/MENA",
    "title": "SAHRA API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/owner/reservations/{id}/cancel": {
      "post": {
        "operationId": "OwnerReservationActionsController_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelledReservationResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "validation_failed — a reason is REQUIRED"
          },
          "403": {
            "description": "not_an_owner"
          },
          "404": {
            "description": "reservation_not_found — ALSO returned for another venue's reservation, deliberately, and byte-identical to an id that exists for nobody."
          },
          "409": {
            "description": "invalid_status_transition — already settled"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a booking, as the restaurant",
        "tags": [
          "owner:reservations"
        ]
      }
    },
...
    "/v1/reservations": {
      "get": {
        "operationId": "MyReservationsController_listMyReservations",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "upcoming",
                "past"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyReservationResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "invalid_query_param"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The caller's own reservations",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/v1/reservations/{id}": {
      "get": {
        "operationId": "MyReservationsController_one",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "reservation_not_found — ALSO returned for a reservation belonging to someone else, deliberately. A 403 would confirm the row exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One of the caller's own reservations",
        "tags": [
          "reservations"
        ]
      },
      "patch": {
        "operationId": "MyReservationsController_modify",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "validation_failed — no fields, or a past time"
          },
          "404": {
            "description": "reservation_not_found — ALSO returned for another diner's reservation, deliberately, and byte-identical to an id that exists for nobody."
          },
          "409": {
            "description": "slot_taken | pacing_limit_reached | invalid_status_transition | reservation_not_modifiable (already started)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change the time or party size of your own booking",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/v1/reservations/{id}/cancel": {
      "post": {
        "operationId": "MyReservationsController_cancelOwn",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOwnReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReservationResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "reservation_not_found"
          },
          "409": {
            "description": "invalid_status_transition — already settled"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel your own booking",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/v1/saved": {
      "get": {
        "operationId": "FavoritesController_listSaved",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedVenueResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The caller's saved venues, newest first",
        "tags": [
          "saved"
        ]
      },
      "post": {
        "operationId": "FavoritesController_save",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveVenueDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already saved"
          },
          "201": {
            "description": "Saved"
          },
          "404": {
            "description": "restaurant_not_found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save a venue (idempotent)",
        "tags": [
          "saved"
        ]
      }
    },
    "/v1/saved/{restaurantId}": {
      "delete": {
        "operationId": "FavoritesController_unsave",
        "parameters": [
          {
            "in": "path",
            "name": "restaurantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Not saved any more, either way"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unsave a venue (idempotent)",
        "tags": [
          "saved"
        ]
      }
    },
    "/v1/waitlists": {
      "get": {
        "operationId": "WaitlistController_listWaitlists",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WaitlistEntryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The caller's live waitlist entries",
        "tags": [
          "waitlists"
        ]
      },
      "post": {
        "operationId": "WaitlistController_join",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinWaitlistDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistEntryResponse"
                }
              }
            },
            "description": ""
          },
          "201": {
            "description": "Joined"
          },
          "400": {
            "description": "validation_failed — past date, or window inverted"
          },
          "404": {
            "description": "restaurant_not_found"
          },
          "409": {
            "description": "already_on_waitlist"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join the waitlist for a date and window",
        "tags": [
          "waitlists"
        ]
      }
    },
    "/v1/waitlists/{id}": {
      "delete": {
        "operationId": "WaitlistController_leave",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Left (the row is kept, marked cancelled)"
          },
          "404": {
            "description": "waitlist_entry_not_found — ALSO for another diner's entry, deliberately."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave the waitlist",
        "tags": [
          "waitlists"
        ]
      }
    }
  },
  "servers": [],
...
}